// src/api/generated/model/recipeDetailResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Bistro App Backend API
 * OpenAPI contract for generated frontend API clients.
 * OpenAPI spec version: 0.1.0
 */
import type { RecipeDetail } from './recipeDetail';

export interface RecipeDetailResponse {
  recipe: RecipeDetail;
}