// src/api/generated/model/getIngredientsScope.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Bistro App Backend API
 * OpenAPI contract for generated frontend API clients.
 * OpenAPI spec version: 0.1.0
 */

export type GetIngredientsScope = typeof GetIngredientsScope[keyof typeof GetIngredientsScope];


export const GetIngredientsScope = {
  all: 'all',
  global: 'global',
  household: 'household',
} as const;