// src/api/generated/model/ingredientBaseUnit.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Bistro App Backend API
 * OpenAPI contract for generated frontend API clients.
 * OpenAPI spec version: 0.1.0
 */

export type IngredientBaseUnit = typeof IngredientBaseUnit[keyof typeof IngredientBaseUnit];


export const IngredientBaseUnit = {
  g: 'g',
  ml: 'ml',
} as const;